update fuelphp to 1.2.1

+++ fuel/core/classes/database/pdo/connection.php
@@ -83,13 +83,22 @@ class Database_PDO_Connection extends \Database_Connection
 		}
 		catch (\PDOException $e)
 		{
-			throw new \Database_Exception($e->getMessage(), $e->getCode(), $e);
+			$error_code = is_numeric($e->getCode()) ? $e->getCode() : 0;
+			throw new \Database_Exception($e->getMessage(), $error_code, $e);
 		}
 
 		if ( ! empty($this->_config['charset']))
 		{
-			// Set the character set
-			$this->set_charset($this->_config['charset']);
+			// Set Charset for SQL Server connection
+			if (strtolower($this->driver_name()) == 'sqlsrv')
+			{
+				$this->_connection->setAttribute(\PDO::SQLSRV_ATTR_ENCODING, \PDO::SQLSRV_ENCODING_SYSTEM);
+			}
+			else
+			{
+				// Set the character set
+				$this->set_charset($this->_config['charset']);
+			}
 		}
 	}
 
@@ -99,6 +108,15 @@ class Database_PDO_Connection extends \Database_Connection
 		$this->_connection = null;
 
 		return true;
+	}
+
+	/**
+	 * Get the current PDO Driver name
+	 * @return string
+	 */
+	public function driver_name()
+	{
+		return $this->_connection->getAttribute(\PDO::ATTR_DRIVER_NAME);
 	}
 
 	public function set_charset($charset)
@@ -121,21 +139,36 @@ class Database_PDO_Connection extends \Database_Connection
 			$benchmark = \Profiler::start("Database ({$this->_instance})", $sql);
 		}
 
-		try
-		{
-			$result = $this->_connection->query($sql);
-		}
-		catch (\Exception $e)
-		{
-			if (isset($benchmark))
-			{
-				// This benchmark is worthless
-				\Profiler::delete($benchmark);
-			}
+		// run the query. if the connection is lost, try 3 times to reconnect
+		$attempts = 3;
 
-			// Convert the exception in a database exception
-			throw new \Database_Exception($e->getMessage().' with query: "'.$sql.'"');
+		do
+		{
+			try
+			{
+				$result = $this->_connection->query($sql);
+				break;
+			}
+			catch (\Exception $e)
+			{
+				if (strpos($e->getMessage(), '2006 MySQL') !== false)
+				{
+					$this->connect();
+				}
+				else
+				{
+					if (isset($benchmark))
+					{
+						// This benchmark is worthless
+						\Profiler::delete($benchmark);
+					}
+
+					// Convert the exception in a database exception
+					throw new \Database_Exception($e->getMessage().' with query: "'.$sql.'"');
+				}
+			}
 		}
+		while ($attempts-- > 0);
 
 		if (isset($benchmark))
 		{
@@ -189,23 +222,23 @@ class Database_PDO_Connection extends \Database_Connection
 	public function list_columns($table, $like = null)
 	{
 		$this->_connection or $this->connect();
-		$q = $this->_connection->prepare("DESCRIBE " . $table);
+		$q = $this->_connection->prepare("DESCRIBE ".$table);
 		$q->execute();
-		$result = $q->fetchAll();
-		$count = 0;
+		$result  = $q->fetchAll();
+		$count   = 0;
 		$columns = array();
-		is_null($like) and str_replace('%', '.*', $like);
+		! is_null($like) and $like = str_replace('%', '.*', $like);
 		foreach ($result as $row)
 		{
-			if (!is_null($like) and preg_match($like, $row['Field'])) continue;
+			if ( ! is_null($like) and ! preg_match('#'.$like.'#', $row['Field'])) continue;
 			list($type, $length) = $this->_parse_type($row['Type']);
 
 			$column = $this->datatype($type);
 
-			$column['name'] = $row['Field'];
-			$column['default'] = $row['Default'];
-			$column['data_type'] = $type;
-			$column['null'] = ($row['Null'] == 'YES');
+			$column['name']             = $row['Field'];
+			$column['default']          = $row['Default'];
+			$column['data_type']        = $type;
+			$column['null']             = ($row['Null'] == 'YES');
 			$column['ordinal_position'] = ++$count;
 			switch ($column['type'])
 			{
@@ -237,28 +270,37 @@ class Database_PDO_Connection extends \Database_Connection
 						case 'tinytext':
 						case 'mediumtext':
 						case 'longtext':
-							$column['collation_name'] = @$row['Collation'];
+							$column['collation_name'] = isset($row['Collation']) ? $row['Collation'] : null;
 							break;
 
 						case 'enum':
 						case 'set':
-							$column['collation_name'] = @$row['Collation'];
-							$column['options'] = explode('\',\'', substr($length, 1, -1));
+							$column['collation_name'] = isset($row['Collation']) ? $row['Collation'] : null;
+							$column['options']        = explode('\',\'', substr($length, 1, - 1));
 							break;
 					}
 					break;
 			}
 
 			// MySQL attributes
-			$column['comment'] = @$row['Comment'];
-			$column['extra'] = $row['Extra'];
-			$column['key'] = $row['Key'];
-			$column['privileges'] = @ $row['Privileges'];
+			$column['comment']    = isset($row['Comment']) ? $row['Comment'] : null;
+			$column['extra']      = $row['Extra'];
+			$column['key']        = $row['Key'];
+			$column['privileges'] = isset($row['Privileges']) ? $row['Privileges'] : null;
 
 			$columns[$row['Field']] = $column;
 		}
 
 		return $columns;
+	}
+
+	public function datatype($type)
+	{
+		// try to determine the datatype
+		$datatype = parent::datatype($type);
+
+		// if not an ANSI database, assume it's string
+		return empty($datatype) ? array('type' => 'string') : $datatype;
 	}
 
 	public function escape($value)
